scripts/requirements.txt
bitarray==3.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
pycparser==2.22
PyWavefront
